Feat: Structure project
