--- lib/src/core/translations/l10n/intl_es.arb
@@ -1,1 +1,0 @@
-{}
