+++ lib/src/core/utils/functions.dart
@@ -1,5 +1,4 @@
 import 'package:url_launcher/url_launcher.dart';
-import 'package:protiendas/app/models/archetype.dart';
 import 'package:protiendas/src/shared/models/list_yugioh.dart';
 
 class Functions {
@@ -10,18 +9,6 @@ class Functions {
     )) {
       throw Exception('Could not launch $url');
     }
-  }
-
-  static List<Archetype> getListFilter({
-    required List<Archetype> listArchetype,
-    required Archetype search,
-  }) {
-    final listFilter = listArchetype
-        .where(
-          (element) => element.archetypeName.contains(search.archetypeName),
-        )
-        .toList();
-    return listFilter;
   }
 
   static List<YuGiOh> getListWithOutBan({
